Work around a flaky Android-23 crash with annotation proxies.
This CL addresses the problem only for @TestParameter itself. However, I expect that to address nearly all problems that people see in practice.

The problem seems to be that Android generates a broken runtime implementation of TestParameterAnnotation. That implementation sometimes crashes when we try to access elements of the annotation, such as valueProvider.

As a workaround, we avoid accessing those elements when we're looking at @TestParameter: We already know what values it uses for TestParameterAnnotation, so we can behave accordingly without having to look them up in the annotation at runtime.